add decimal functionality
logicaltype setup

fix

fix

changes

added more switch handling for decimal types

more support

compile fix

function work

stub

work

reconfigure cypher_lexer

Run clang-format

clang-tidy fix

clang tidy

Run clang-format

clang-tidy

add requested changes

requested changes

more requested changes

fix

Run clang-format

add requested changes

fustfmt

update parser and lexer

add prelim changes

logicaltype setup

fix

fix

changes

added more switch handling for decimal types

more support

compile fix

function work

stub

work

reconfigure cypher_lexer

Run clang-format

clang-tidy fix

clang tidy

Run clang-format

clang-tidy

add requested changes

requested changes

more requested changes

fix

Run clang-format

add requested changes

fustfmt

arithmetic functions

progress

work

progress

actually register the arithmetic functions...

remove cast decimal

progress

finished(?)

roundign

progress

now it nearly compiles!

compiles

add test

test fix

implicit cast addendum

add decimal support for maxlogicaltype

touch ups

casting overflow check

remove lasttest log

add typename for using clauses

fix floor/ceil

clang tidy

fix compile error

Run clang-format

clang-tidy

fix tests

cast tests

fix binder error test

Run clang-format

add requested changes

CI

Run clang-format

fix msvc build cI

CI

ci fix

fix casting bug

Run clang-format

compiler fix

Run clang-format

compiler fix 2

compiler issue fix?

test

test to see

fix (hopefully)

Run clang-format

compiler fix

oops

add requested tests

Run clang-format

add requested changes 2

CI fix

remove mm-256kb

fix bug

requested changes

clang tidy

Run clang-format

+++ test/main/prepare_test.cpp
@@ -230,7 +230,7 @@ TEST_F(ApiTest, MultipleExecutionOfPreparedStatement) {
 TEST_F(ApiTest, issueTest4) {
     auto preparedStatement = conn->prepare("RETURN CAST($1, 'STRING')");
     auto result = conn->execute(preparedStatement.get(),
-        std::make_pair(std::string("1"), int128_t(-123456789)));
+        std::make_pair(std::string("1"), int128_t((int32_t)-123456789)));
     ASSERT_TRUE(result->hasNext());
     checkTuple(result->getNext().get(), "-123456789\n");
     ASSERT_FALSE(result->hasNext());
